refactor: move ticket-create helpers out of commands/ (fix layer violation)
/audit (depcruise cli-no-cross-command-imports) flagged the orchestration
helpers I'd placed under commands/: commands/ is for entry points only and
must not cross-import. Move them into a feature subdir (matches tracker-sync/
ticket-sync/ convention):

  commands/create-ticket-routed.ts → ticket-create/index.ts
  commands/ticket-creation-mode.ts → ticket-create/creation-mode.ts
  commands/ticket-identity.ts      → ticket-create/identity.ts

commands/ticket-new.ts (the real command) stays and imports from
../ticket-create. Pure move + import-path updates; depcruise clean (0
violations), typecheck green, 352 passed / 4 skipped.
